car_wash_service/requirements.txt
fastapi
uvicorn
motor
pydantic
pyjwt
python-dotenv
httpx
